Refactor : form 컴포넌트 분기처리
