+++ src/components/InputForm/Form.module.scss
@@ -24,13 +24,24 @@
       padding-left: 5px;
     }
   }
+ //address form
 
-  //address form
-  .zonecode{
-    input{
-      width: 32%;
+  .address{
+    .zonecode{
+      display: flex;
+      input{
+        display: inline-block;
+        width: 32%;
+      }
+      button{
+        display: inline-block;
+        flex-shrink: 0;
+        margin-left: 10px;
+      }
     }
   }
+
+ 
 
   //email form
   .inputEmail{
