+++ src/pages/JoinUs/JoinUs.module.scss
@@ -38,16 +38,7 @@
           width: 100%;
           font-size: .9rem;
         }
-        .address{
-          input{
-            display: inline-block;
-          }
-          button{
-            display: inline-block;
-            flex-shrink: 0;
-            margin-left: 10px;
-          }
-        }
+      
       }
     }
     .agreeWrap{
